:scissors: move interfaces from supervisor file
